refactor(network-segment-controller): extract into dedicated crate
Move the network segment state controller out of `crates/api` into a
new `carbide-network-segment-controller` crate as part of the ongoing
controller extraction work.
