+++ crates/api/Cargo.toml
@@ -46,6 +46,7 @@ carbide-ib-fabric = { path = "../ib-fabric" }
 carbide-ipxe-renderer = { path = "../ipxe-renderer" }
 carbide-ipmi = { path = "../ipmi" }
 carbide-redfish = { path = "../redfish" }
+carbide-network-segment-controller = { path = "../network-segment-controller" }
 carbide-site-explorer = { path = "../site-explorer" }
 carbide-spdm-controller = { path = "../spdm-controller" }
 carbide-preingestion-manager = { path = "../preingestion-manager" }
