+++ crates/api/src/state_controller/mod.rs
@@ -21,7 +21,6 @@ pub(crate) mod external_service_error;
 pub mod health_metrics;
 pub mod ib_partition;
 pub mod machine;
-pub mod network_segment;
 pub mod power_shelf;
 pub mod rack;
 pub mod switch;
